o Move tests around and add docs
